Small refactor of cerner drupal-source-of-truth selectors.
