+++ src/applications/static-pages/health-care-manage-benefits/refill-track-prescriptions-page/components/App/index.js
@@ -5,10 +5,8 @@ import { connect } from 'react-redux';
 import { isEmpty } from 'lodash';
 // Relative imports.
 import { isAuthenticatedWithSSOe } from 'platform/user/authentication/selectors';
-import {
-  selectPatientFacilities,
-  selectPatientFacilitiesDsot,
-} from 'platform/user/selectors';
+import { selectPatientFacilities } from 'platform/user/selectors';
+import { selectPatientFacilities as selectPatientFacilitiesDsot } from 'platform/user/cerner-dsot/selectors';
 import AuthContent from '../AuthContent';
 import UnauthContent from '../UnauthContent';
