+++ src/applications/static-pages/health-care-manage-benefits/schedule-view-va-appointments-page/components/App/index.js
@@ -4,10 +4,8 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { isEmpty } from 'lodash';
 // Relative imports.
-import {
-  selectPatientFacilities,
-  selectPatientFacilitiesDsot,
-} from 'platform/user/selectors';
+import { selectPatientFacilities } from 'platform/user/selectors';
+import { selectPatientFacilities as selectPatientFacilitiesDsot } from 'platform/user/cerner-dsot/selectors';
 import AuthContent from '../AuthContent';
 import UnauthContent from '../UnauthContent';
