+++ src/platform/user/tests/selectors.unit.spec.js
@@ -377,254 +377,6 @@ describe('user selectors', () => {
     });
   });
 
-  describe('selectCernerFacilitiesDsot', () => {
-    it('pulls out state.drupalStaticData.cernerFacilities.data when set on state', () => {
-      const state = {
-        drupalStaticData: {
-          cernerFacilities: {
-            data: [
-              {
-                vhaId: '757',
-                vamcFacilityName:
-                  'Chalmers P. Wylie Veterans Outpatient Clinic',
-                vamcSystemName: 'VA Central Ohio health care',
-                ehr: 'cerner',
-              },
-              {
-                vhaId: '687',
-                vamcFacilityName:
-                  'Jonathan M. Wainwright Memorial VA Medical Center',
-                vamcSystemName: 'VA Walla Walla health care',
-                ehr: 'cerner',
-              },
-            ],
-          },
-        },
-      };
-      expect(selectors.selectCernerFacilitiesDsot(state)).to.deep.equal(
-        state.drupalStaticData.cernerFacilities.data,
-      );
-    });
-    it('returns empty array when Cerner facilities not set on state', () => {
-      const state = {};
-      expect(selectors.selectCernerFacilitiesDsot(state)).to.deep.equal([]);
-    });
-  });
-
-  describe('selectCernerFacilityIdsDsot', () => {
-    it('returns non-empty, flat array of ids when Cerner facilities set on state', () => {
-      const state = {
-        drupalStaticData: {
-          cernerFacilities: {
-            data: [
-              {
-                vhaId: '757',
-                vamcFacilityName:
-                  'Chalmers P. Wylie Veterans Outpatient Clinic',
-                vamcSystemName: 'VA Central Ohio health care',
-                ehr: 'cerner',
-              },
-              {
-                vhaId: '687',
-                vamcFacilityName:
-                  'Jonathan M. Wainwright Memorial VA Medical Center',
-                vamcSystemName: 'VA Walla Walla health care',
-                ehr: 'cerner',
-              },
-            ],
-          },
-        },
-      };
-      expect(selectors.selectCernerFacilityIdsDsot(state)).to.deep.equal([
-        '757',
-        '687',
-      ]);
-    });
-    it('returns empty array when Cerner facilities not set on state', () => {
-      const state = {};
-      expect(selectors.selectCernerFacilityIdsDsot(state)).to.deep.equal([]);
-    });
-  });
-
-  describe('selectPatientFacilitiesDsot', () => {
-    it('returns null when there are no facilities on the profile', () => {
-      const state = {
-        user: {
-          profile: {},
-        },
-      };
-      expect(selectors.selectPatientFacilitiesDsot(state)).to.be.null;
-    });
-    it('pulls out the state.profile.facilities array', () => {
-      const state = {
-        user: {
-          profile: {
-            facilities: [
-              { facilityId: '983', isCerner: false },
-              { facilityId: '984', isCerner: false },
-            ],
-          },
-        },
-      };
-      expect(selectors.selectPatientFacilitiesDsot(state)).to.deep.equal(
-        state.user.profile.facilities,
-      );
-    });
-    it('pulls out state.profile.facilities and marks all isCerner:false when Cerner facilities not set on state', () => {
-      const state = {
-        user: {
-          profile: {
-            facilities: [
-              { facilityId: '984' },
-              { facilityId: '668' },
-              { facilityId: '757' },
-            ],
-          },
-        },
-      };
-      const expected = [
-        { facilityId: '984', isCerner: false },
-        { facilityId: '668', isCerner: false },
-        { facilityId: '757', isCerner: false },
-      ];
-      expect(selectors.selectPatientFacilitiesDsot(state)).to.deep.equal(
-        expected,
-      );
-    });
-    it('pulls out state.profile.facilities and marks isCerner according to Cerner facilities data set on state', () => {
-      const state = {
-        user: {
-          profile: {
-            facilities: [
-              { facilityId: '984' },
-              { facilityId: '687' },
-              { facilityId: '757' },
-            ],
-          },
-        },
-        drupalStaticData: {
-          cernerFacilities: {
-            data: [
-              {
-                vhaId: '757',
-                vamcFacilityName:
-                  'Chalmers P. Wylie Veterans Outpatient Clinic',
-                vamcSystemName: 'VA Central Ohio health care',
-                ehr: 'cerner',
-              },
-              {
-                vhaId: '687',
-                vamcFacilityName:
-                  'Jonathan M. Wainwright Memorial VA Medical Center',
-                vamcSystemName: 'VA Walla Walla health care',
-                ehr: 'cerner',
-              },
-            ],
-          },
-        },
-      };
-      const expected = [
-        { facilityId: '984', isCerner: false },
-        {
-          facilityId: '687',
-          isCerner: true,
-          usesCernerAppointments: true,
-          usesCernerMedicalRecords: true,
-          usesCernerMessaging: true,
-          usesCernerRx: true,
-          usesCernerTestResults: true,
-        },
-        {
-          facilityId: '757',
-          isCerner: true,
-          usesCernerAppointments: true,
-          usesCernerMedicalRecords: true,
-          usesCernerMessaging: true,
-          usesCernerRx: true,
-          usesCernerTestResults: true,
-        },
-      ];
-      expect(selectors.selectPatientFacilitiesDsot(state)).to.deep.equal(
-        expected,
-      );
-    });
-  });
-
-  describe('selectPatientCernerFacilitiesDsot', () => {
-    it('returns empty array when Cerner facilities not set on state', () => {
-      const state = {
-        user: {
-          profile: {
-            facilities: [
-              { facilityId: '984' },
-              { facilityId: '687' },
-              { facilityId: '757' },
-            ],
-          },
-        },
-      };
-      expect(selectors.selectPatientCernerFacilitiesDsot(state)).to.deep.equal(
-        [],
-      );
-    });
-    it('pulls out patient facilities that are indicated as Cerner facilities', () => {
-      const state = {
-        user: {
-          profile: {
-            facilities: [
-              { facilityId: '984' },
-              { facilityId: '687' },
-              { facilityId: '757' },
-            ],
-          },
-        },
-        drupalStaticData: {
-          cernerFacilities: {
-            data: [
-              {
-                vhaId: '757',
-                vamcFacilityName:
-                  'Chalmers P. Wylie Veterans Outpatient Clinic',
-                vamcSystemName: 'VA Central Ohio health care',
-                ehr: 'cerner',
-              },
-              {
-                vhaId: '687',
-                vamcFacilityName:
-                  'Jonathan M. Wainwright Memorial VA Medical Center',
-                vamcSystemName: 'VA Walla Walla health care',
-                ehr: 'cerner',
-              },
-            ],
-          },
-        },
-      };
-      const expected = [
-        {
-          facilityId: '687',
-          isCerner: true,
-          usesCernerAppointments: true,
-          usesCernerMedicalRecords: true,
-          usesCernerMessaging: true,
-          usesCernerRx: true,
-          usesCernerTestResults: true,
-        },
-        {
-          facilityId: '757',
-          isCerner: true,
-          usesCernerAppointments: true,
-          usesCernerMedicalRecords: true,
-          usesCernerMessaging: true,
-          usesCernerRx: true,
-          usesCernerTestResults: true,
-        },
-      ];
-      expect(selectors.selectPatientCernerFacilitiesDsot(state)).to.deep.equal(
-        expected,
-      );
-    });
-  });
-
   describe('selectIsCernerOnlyPatient', () => {
     it('should return true if Cerner only', () => {
       const state = {
@@ -658,68 +410,6 @@ describe('user selectors', () => {
         },
       };
       expect(selectors.selectIsCernerOnlyPatient(state)).to.be.false;
-    });
-  });
-  describe('selectIsCernerOnlyPatientDsot', () => {
-    it('returns true when all patient facilities are indicated as Cerner facilities', () => {
-      const state = {
-        user: {
-          profile: {
-            facilities: [{ facilityId: '757' }, { facilityId: '687' }],
-          },
-        },
-        drupalStaticData: {
-          cernerFacilities: {
-            data: [
-              {
-                vhaId: '757',
-                vamcFacilityName:
-                  'Chalmers P. Wylie Veterans Outpatient Clinic',
-                vamcSystemName: 'VA Central Ohio health care',
-                ehr: 'cerner',
-              },
-              {
-                vhaId: '687',
-                vamcFacilityName:
-                  'Jonathan M. Wainwright Memorial VA Medical Center',
-                vamcSystemName: 'VA Walla Walla health care',
-                ehr: 'cerner',
-              },
-            ],
-          },
-        },
-      };
-      expect(selectors.selectIsCernerOnlyPatientDsot(state)).to.be.true;
-    });
-    it('returns false when at least one patient facility is not a Cerner facility', () => {
-      const state = {
-        user: {
-          profile: {
-            facilities: [{ facilityId: '757' }, { facilityId: '123' }],
-          },
-        },
-        drupalStaticData: {
-          cernerFacilities: {
-            data: [
-              {
-                vhaId: '757',
-                vamcFacilityName:
-                  'Chalmers P. Wylie Veterans Outpatient Clinic',
-                vamcSystemName: 'VA Central Ohio health care',
-                ehr: 'cerner',
-              },
-              {
-                vhaId: '687',
-                vamcFacilityName:
-                  'Jonathan M. Wainwright Memorial VA Medical Center',
-                vamcSystemName: 'VA Walla Walla health care',
-                ehr: 'cerner',
-              },
-            ],
-          },
-        },
-      };
-      expect(selectors.selectIsCernerOnlyPatientDsot(state)).to.be.false;
     });
   });
 
@@ -771,99 +461,6 @@ describe('user selectors', () => {
         },
       };
       expect(selectors.selectIsCernerPatient(state)).to.be.false;
-    });
-  });
-
-  describe('selectIsCernerPatientDsot', () => {
-    it('returns true when single patient facility is indicated as a Cerner facility', () => {
-      const state = {
-        user: {
-          profile: {
-            facilities: [{ facilityId: '757' }],
-          },
-        },
-        drupalStaticData: {
-          cernerFacilities: {
-            data: [
-              {
-                vhaId: '757',
-                vamcFacilityName:
-                  'Chalmers P. Wylie Veterans Outpatient Clinic',
-                vamcSystemName: 'VA Central Ohio health care',
-                ehr: 'cerner',
-              },
-              {
-                vhaId: '687',
-                vamcFacilityName:
-                  'Jonathan M. Wainwright Memorial VA Medical Center',
-                vamcSystemName: 'VA Walla Walla health care',
-                ehr: 'cerner',
-              },
-            ],
-          },
-        },
-      };
-      expect(selectors.selectIsCernerPatientDsot(state)).to.be.true;
-    });
-    it('returns true when at least one patient facility is indicated as a Cerner facility', () => {
-      const state = {
-        user: {
-          profile: {
-            facilities: [{ facilityId: '757' }, { facilityId: '123' }],
-          },
-        },
-        drupalStaticData: {
-          cernerFacilities: {
-            data: [
-              {
-                vhaId: '757',
-                vamcFacilityName:
-                  'Chalmers P. Wylie Veterans Outpatient Clinic',
-                vamcSystemName: 'VA Central Ohio health care',
-                ehr: 'cerner',
-              },
-              {
-                vhaId: '687',
-                vamcFacilityName:
-                  'Jonathan M. Wainwright Memorial VA Medical Center',
-                vamcSystemName: 'VA Walla Walla health care',
-                ehr: 'cerner',
-              },
-            ],
-          },
-        },
-      };
-      expect(selectors.selectIsCernerPatientDsot(state)).to.be.true;
-    });
-    it('returns false when no patient facilities are indicated as Cerner facilities', () => {
-      const state = {
-        user: {
-          profile: {
-            facilities: [{ facilityId: '123' }, { facilityId: '124' }],
-          },
-        },
-        drupalStaticData: {
-          cernerFacilities: {
-            data: [
-              {
-                vhaId: '757',
-                vamcFacilityName:
-                  'Chalmers P. Wylie Veterans Outpatient Clinic',
-                vamcSystemName: 'VA Central Ohio health care',
-                ehr: 'cerner',
-              },
-              {
-                vhaId: '687',
-                vamcFacilityName:
-                  'Jonathan M. Wainwright Memorial VA Medical Center',
-                vamcSystemName: 'VA Walla Walla health care',
-                ehr: 'cerner',
-              },
-            ],
-          },
-        },
-      };
-      expect(selectors.selectIsCernerPatientDsot(state)).to.be.false;
     });
   });
